[kernel][vm] Implement clone for pager vmos
Implement COW clones of pager vmos. These clones behave like private
posix mappings (i.e. writes are not visible to other clients of the
original pager vmo and are not propagated back to the pager).

This changes how clones get pages from their parent. Instead of masking
out the pf page fault bits, it masks out the pf write bit. This is
necessary because a fault on the child does actually need to generate a
read fault on the parent.

Test: /boot/test/core/pager-test
Test: Pave, boot, use session shell on eve
Test: vmstress from sandbox/stevensd/userpager for 48+ hours on
nuc/vim
Test: repeatedly run and kill vmstress for 12+ hours on nuc

Change-Id: Ieb6986c6fefd5fa3450d2affefba49c8eceb1e8c