[multiboot] Make the Multiboot trampoline loader compatible with CrOS
The CrOS boot loader supports Multiboot with special features when
the RAM disk payload is in ZBI format.  Its implementation of this
violates three assumptions the Multiboot trampoline loader made.

Firstly, it prepends items to the ZBI so that the kernel is no longer
guaranteed to be the first item as in a complete ZBI as built.

Secondly, it fails to update the Multiboot headers to adjust the size
of the Multiboot module payload that holds the ZBI.  This means that
a sanity check of the ZBI container header against the Multiboot module
size cannot be done.

Thirdly, it provides Multiboot MMAP information but not MEM_SIZE (as qemu
does).  So determining the upper limit of available contiguous physical
memory requires a little more work.

Bug: BLD-269 #comment Multiboot trampoline compatible with CrOS bootloader
Test: CQ (tests qemu) and manual paving/booting of Pixelbook using multiboot.bin as kernel and ZBI as ramdisk
Change-Id: Ice9cecca9f8afceef5228447424ae535b54a96e4