[botanist] Move SSH utility functions to botanist package
Go's ssh package is already sufficient for common use cases, we
shouldn't try to reimplement it with a different interface. Rather,
we only want to provide a set of utility functions for use in botanist.
Move these functions from ssh package to botanist package to make this
clear. This also avoid the collission between different ssh packages.

Change-Id: Idc53d542210a5ac6cfc5e41406d27f571c5c6b40