[benchmarks] add benchmarks for static slab allocator
Add a few slab allocation benchmarks. The benchmarks exercise
multiple allocation sizes (64 bytes and 8KB, respectively), and
multiple allocation patterns.

The motivation for multiple sizes is to quantify any scaling behavior
with the size of the allocation.

The motivation for multiple patterns is to identify the impact of
access patterns on allocation cost:
* The AllocAndFree pattern represents (presumed) best-case behavior,
  as the memory pool should be unfragmented.
* The RetainAndFreeOldest pattern abstracts a network copy workload,
  when copying from a fast source, to a slow sink.
* The RetainAndFreeRandom pattern attempts to quantify the effects
  of memory fragmentation.

Bug: ZX-3242
Test: /pkgfs/packages/zircon_benchmarks/0/test/zircon_benchmarks \
        -p --filter "SlabAlloc/Static/" # on eve

Change-Id: Id79fbcae592e737cd0917e23e92efd0be8f45b18