[amber] Set response timeout
Set a 15 second timeout for getting the first bit of the header
from the server. This is a workaround for not detecting that
cache connections are closed and therefore waiting indefinitely
for a response. This has the potential to do the wrong thing, but
for our purposes right now we expect all reasonably well-behaved
environments to get a response during this time.

TEST=Manual, get one update, kill the devhost update server, try
to get an update again. Before this change the second request
would hang indefinitely, after it eventually times out.

PKG-400 #done

Change-Id: I585e18fa21bbace77cc72f77d7451df360f27330