[fidl] Go-based code generator for the low-level C++ binding.
This is step one of having a client-side API generated to be able to
interop test against a C server. Currently, the relevant FIDL message
definitions are generated, but not the API implementation which
linearizes and encodes/decodes (they just return ZX_ERR_NOT_SUPPORTED).
The next step is to generate these APIs for the case of simple messages.

FIDL-377 #done

TEST: The generated code is tested in Zircon, via
https://fuchsia-review.googlesource.com/c/zircon/+/232379

Change-Id: Ida7f8bae047bc2022623b81eb0fe03bf75fae938