[fidl] Refactor C++ codegen flow to be consistent with the others.
Dart/Rust/Go code generators emit type definitions in the same order as
that from the JSON IR. However, due to the limitations of the C++
language, type definitions need to be emitted in "declaration order",
which is a topologically sorted order by type dependencies.

This introduced an extra layer of indirection in the C++ codegen, making
it a bit hard to reason about the flow in the templates.

This change attempts to simplify code generation in C++ by having
templates call type-specific sub-templates, as opposed to a dyamic
dispatch logic outside of go templates. It does not affect generated code.
They will remain completely identical.

More context: Pascal's comments in
https://fuchsia-review.googlesource.com/c/garnet/+/230898/23

Test: fx run-host-tests fidlgen_cpp_test && fx run-host-tests
fidlgen_llcpp_test

Change-Id: Ib5ad7c25ddfd27ee02fabbeb8067b31a3b2edd9c