[bt][gap] LE local address management for privacy
Introduce the gap::LowEnergyAddressManager which manages
the local controller address used for scan, connection initiation, and
legacy advertising procedures. Storing the state of the local address
in a central location is necessary since the controller address is
shared by these procedures and modifying it requires care when different
procedures are active concurrently.

Bug: BT-611
Test: bt-host-unittests
Change-Id: Iea146da058f31f823cca53d4bad159e3303c10ed