[auth] Move AuthProviderClient into TokenManager lib
The previous refactoring CL left the launching of auth providers in
TokenManagerFactory, but AccountManager will also need to perform the
same functionality and so its probably better in the reusable lib.

Doing this as a separate CL for clarity, and renaming to ...Connection
for consistency with the similar AccountHandlerConnection.

Tested: CQ, specifically token manager integration test.
Change-Id: I08fc7196841667fe8fcf40d9835469b85111a9de