[tracing][benchmark] Add more macros to trace-benchmark
The existing benchmark relied mostly on TRACE_DURATION_BEGIN.
Most macros had almost the same logic and the shared the same code,
but as we explore more complicated macros, it is worth adding more
macros to the performance test.

This CL generalizes the test for any duration macro, and it is
currently tested with the original TRACE_DURATION_BEGIN as well as
TRACE_DURATION.

TEST: fx shell trace-benchmark
Verified that the original output is unchanged as well as the addition
of the new macro.

Change-Id: I3061a72237034246bf904e73cbb49a0b232cbaf0