[hypervisor][arm64] Use caller-saved registers
When storing arguments for HVC invocations, ensure we use caller-saved
registers.

Since we control the ABI between EL1 and EL2, this change gives us some
additional flexibility with the number of arguments that can be passed
to these functions without trampling on the additional arguments that
need to be set for EL2 to decode the HVC invocation.

Test: Ran "linux_guest" on a VIM2 and QEMU.
Change-Id: Ie9ca97674f90f4aefb661d90a6ad4c4d6469cd7a