[auth] Add context to AccountHandler init calls.
Previously I'd been working on the basis that we'd be able to insert the
AccountHandlerContext directly into the AccountHandler's namespace while
launching it, but a recent email thread indicates this isn't going to
work well until the next version of component architecture. Until then
I'm moving to explicitly supplying the context in the calls that are
used to initialize AccountHandler.

AUTH-107

Tested: CQ, demonstrates no breakage. This CL doesn't yet add any new
functionality to test.

Change-Id: I2901d1dfc235ef89b3f29c99518aa5ef37e28a6f