[fidl] Visitor-style message linearizing.
Linearize an object tree, driven by the corresponding coding table.
This paves way to generating client/server interfaces in LLCPP.
When the user wants to send a FIDL message that is completely inline,
linearization can be skipped, and the message goes directly to encoding.
When the message contains pointers however, the FIDL wireformat
specifies that the pointee must occurr in depth-first traversal order.
Hence the need arises to collate (linearize) the objects.

The C binding produces a very limited linearization implementation
which only works for structures with up to one level of indirection.
By leveraging coding tables, we can perform linearization on arbitrarily
nested objects if necessary.

As a bonus, handle safety is enforced by clearing out any handles in
the original objects, effectively moving them to the linearized buffer.

FIDL-417 #done

TEST: /boot/test/sys/fidl-test

Change-Id: I29b416f77f2ad916a76fc0f68c2abada127ea39c