[botanist] Fix arm64 paving bug.
new(Foo) != nil
new returns a pointer to a Foo struct that is 'zeroed'; this is not
a nil pointer.

This fixes a bug in the arm64 codepath (i.e., for vim2s) in which the
following was happening:
-images.json for an arm64 build did not designate a kernel (with a
--boot switch) (WAI)
-in Pave(), in iterating over the images looking for a kernel, it tries
to reassign a pointer defined as new(Image); because no kernel was found
it passes this non-nil-but-garbage pointer to transfer() which fails in
trying to open it.

Change-Id: I2d1e6eec4f3103bd30b52c932b836206c769ab33