[testing] EnvironmentServices stdout/err passing
EnvironmentServices was only passing url and arguments for launch
services' LaunchInfo (somewhat sneakily). This adds a new method to
EnvironmentServices that allows providing a handler function that acts
as a factory of LaunchInfo objects.

Also adds a test package with some unit tests for lib component.

TEST: ran tests that use enclosing environment and services, added new
tests package to lib component cpp.

Change-Id: I31c18a7c64df4fa412385e5f0fb874bd43bede2b