[brcmfmac] Fix for 4359 firmware-startup bug
The DHD driver doesn't reset the 80211 core prior to firmware load. The
brcmfmac driver does. On some 4359 chips, some of the time, this causes
an error on firmware boot.

Note: This code does not properly handle driver unload, when the 80211
core should presumably be disabled.

WLAN-745 #done
Test: Verified no regression on 4356 with Belkin, Netgear, and TP-Link
routers. Verified on 4359 with ASUS, Belkin, Netgear, and TP-Link routers.

Change-Id: If6e28798f73b2e7dc09dd6203f465b5f0da0e3d9