[mdns] fix address fix-up to enable V6
The per-address 'transceivers' in the mDNS service modify DNS
messages before sending them in order to replace placeholder address
(A) resources with the correct address resources for that interface.
This includes an address record for the interface itself (A for a V4
interface, AAAA for a V6 interface). It may also include a resource
for the 'alternate' address, that is, the alternate Vx address for
the same NIC.

Because the same message is reused sequentially across multiple
transceivers, a given transceiver may encounter an already-modified
message when doing its address fix-ups. The logic for doing fix-ups
didn't allow for this, insisting that the original message should
contain no AAAA records and not removing redundant address records.
This logic has been fixed so that all original A or AAAA records
are removed from a given resource list and collectively replaced
with the correct set of address resources for the transceiver.

https://fuchsia-review.googlesource.com/c/garnet/+/224838 is
required to make this work, so NET-1900 isn't marked done by this
CL.

TEST: boot, run avahi-resolve -6 -n cadet-cake-sugar-whoop.local.

NET-1900 #comment

Change-Id: Id555ca5f9e0334c11b70333b73923686c509db62