[zxtest] Add assertions.
Plain helper class for keeping assertion data together.
Updated internal::TestDriverImpl to keep track of errors generated on OnAssertion.
Updated Reporter to print the error messages. If expect/actual and
*_eval are equal the error message is simplified, this is when:L
ASSERT(var, MyLiteralValue) -> (var, 5) produces:
  Actual: var
Which is: some value
Expected: 5
Which is: 5
In gtest, but we can just collapse these two, when the stringified value
matches the formatted value, removing the unnecessary 'Which is:'.

Added Runner::NotifyAssertion(const Assertion& assertion) as an entry
point for macros for propagating errors. This is done instead of a method
in |Test| because legacy tests, c tests, etc dont have access to any
of these, neither do helper methods, so this will allow detecting errors
in any of those contexts. Plus can be used in conjunction with
ASSERT_NO_FATAL_ERRORS, to return from a test method after a call to a
helper method that might fail(this is the approach taken by gtest).

Alternatives considered include setjmp/longjmp but this do not play well
with C++ destructors, specially because ignores any intermediate scopes,
and C++ exceptions are a big no.

TEST=zxtest-test

Change-Id: I6644b04e0d5be84fd70784ae35cc1d29a93278cf