[ledger] do not automatically merge identical commits
Commits with identical contents should not be automatically merged as
they may be two concurrent changes coming from different devices (eg.
two separate incrementations of a counter). This removes this behavior.
Since the common ancestor algorithm is able to identify commits that
include the exact same set of changes, the intended case is already
covered.

TEST=MergeResolverTest/DoNotMergeIdenticalCommits

Change-Id: I77dc8251a92c64c0490acb54a2a8e7f5255ffdd9