[fdio] Remove race from zxsio_wait_begin_stream
Previously, zxsio_wait_begin_stream would issue a zero-timeout wait to
poll the connection state of the socket. This pattern is racy because
the semantics depend on whether the server has set the connected signal
before or after the zero-timeout wait returns.  This CL removes the race
and relies on the logic in zxsio_wait_end_stream to update the socket
state machine.

The original code was added in a769b0f29013b555f2bd7b9fba06bb9f2c82ef9e
as part of landing the connection state machine in the first place.

Test: No behavior change.
Change-Id: Ib6723fcc796452ea7c909b58860fd3b73e326c4b