[pseudo-fs] Switch Future<Output = {Result<(), Error> => Void}>
Pseudo directory entry futures are never expected to complete.  It is
better to assert that statically by using Void instead of ().

Error reporting seems also unnecessary, it was not used nor in file, nor
in the directory implementatio of the DirectoryEntry trait.  It is
unlikely to be necessary for other implementations in the future.

Test: All unit tests pass.
Change-Id: I37e329ba4ca4a738e47b610caf36349a1a513805