[mdns] use gethostname for faster startup
The mDNS service used to poll the network interfaces on startup
waiting for a valid (not 0.0.0.0, not loopback) interface to show
up. This was because gethostname would return unpredictable values
prior to that point. The interfaces were polled every five seconds.
In addition, the valid address detection was really async pretending
to be synchronous and could therefore add another 5 seconds of delay
to mDNS startup. There was also some unnecessary logic that tried to
disambiguate a common default hostname ("fuchsia") used in the past.
Now we have probing to achieve disambiguation.

gethostname now consistently returns "fuchsia-unset-device-name"
until the actual hostname is known, so mDNS can poll gethostname.
The polling frequency has been increased to once/second. The result
is that mDNS starts more promptly now. It gets the host name before
the NICs are ready, so it's able to start functioning immediately
after Netstack.OnInterfacesChanged fires.

TEST: start mDNS, test manually for regressions, look at the logs

Change-Id: I1da43cf27e7dffa8b952df434b215053ad8b49bb