[tee] Add placeholder field to TEEC Operation Impl
TEE Client API's TEEC_Operation structure requires an implementation
defined struct embedded within it. We currently have nothing to put in
the implementation defined struct, but this commit will add a
placeholder field so that the impl struct is still well defined.
Previously, we were relying on a GNU extension to allow for empty
structs in C.

Test: Compiles
Change-Id: Ide51f4237683a039e0aa817a024673790d509cb1