[fdio] Add size response to fuchsia.io.File.GetVmo
Simply getting the VMO is insufficient now that VMO sizes are rounded
up to the nearest page boundary. Clients also need to know how many
bytes in the VMO are part of the file.

They can obtain this information by getting the attributes of the file,
but that's slow and racy. Instead, in this CL, we return the size of the
file along with the VMO.

Test: Will add tests when updating the client code, which is blocked on
adding server implementations in Go.

Change-Id: I5aa32bfc6b8e066fbc9a2e7ca40ab78796cc78f6