[modular][storage] Don't store ModuleManifest in ModuleData.
We stored ModuleManifest in ModuleData so that we could eventually send
it to the StoryShell. Instead of doing that, we can look it up right
before sending a Module's info to the StoryShell and thus remove a
number of unnecessarily locations depending on ModuleFacetReader.

Now, we always use the ModuleFacetReader to get our ModuleManifest.

TEST=existing

Change-Id: If7043b9663571241e28da7f8efd2a4f365a1f7a1