[debugger] Separate out Stack from Thread.
Separates out the stack data and functions from the Thread object into a
new Stack object. This is to support future additions to stack
handling for inline functions which will complicate the code
substantially.

For simplicity, this attempts to change as little as possible and there
should be no behavior change.

TEST=unit

Change-Id: Ic234a1462c74ead58ffd1a16e4fdf4abf5bace94