[cleo][touch] Add FT6336 HID descriptor
The FT3x27 driver works with the FT6336 touch
device, however the FT3x27 descriptor does not
match the FT6336. Add a descriptor for the FT6336
with the correct coordinate range and change the
driver to load it when required.

Also change "ft3x27" to "focaltouch" in the
driver.

ZX-2788
Test: Inspected descriptor with 'hid parse'.
Test: Confirmed dimensions with 'hid read'.

Change-Id: I2a3dd969013d980e0d7ba1c40bcf44a1a6e92142