[mediasession] Add example uses of Media Session.
This adds an example media player that just plays a sine wave which
shows how to use the Media Session API from the player side.

This adds the shell tool `mediasession_cli_tool` which announces changes
to the active session and events from the active session. This is
helpful when implementing `fuchsia.mediasession.Controller` for Media
Session API.

TEST: fx shell mediasession_cli_tool, fx shell run sine_player

Change-Id: I3e7b70f08ebbb611060fad289a5f781a2f0ba16e