[bt][hci] Do not configure random address in legacy advertiser
The management of the local LE random address is now the responsiblity
of a GAP library object that maintains a consistent address for scan,
initiation, and legacy advertising procedures. The following changes are
made to support the new model:

1. hci::LegacyLowEnergyAdvertiser no longer configures the random
device address. The |address| parameter of StartAdvertising() is
used to determine the advertised "own address type" to configure the
local address of remote-initiated links but is not directly assigned to
the controller. The |address| parameter is still being used to track local
advertisement instances.

2. Added the AllowsRandomAddressChange() method to hci::LowEnergyAdvertiser.
This will be used by gap::Adapter to coordinate random address updates.

Bug: BT-611
Test: bt-host-unittests --gtest_filter=GAP_*Advertising*
      bt-host-unittests --gtest_filter=HCI_*Advert*
      bt-host-unittests

Change-Id: I37597745533e4850459f232590f182f5c84b52e0