[auth] Move AuthUiContext to TokenManagerContext
Prior to this CL, the TokenManager refactor treated the supply of auth
provider channels and UI context channels as equivalent, with an object
implementing a supplier trait being supplied at TokenManager
construction.

However, the UI context is a little different: multiple shells could
potentially request channels to the same underlying token manager
instance but supply their own UI context providers, hence the UI
context provider can't be attached to the token manager at
construction, instead it needs to be modelled in the same fashion as
application URL: a part of the channel-specific context that is passed
in the handle requests call.

Since the supply of UI context comes later (and due to new findings in
AUTH-110) there is less need to provide a layer of abstraction over
the creation of new UI contexts so I've also removed the
AuthContextSupplier trait in favor of using the FIDL provider
interface directly.

AUTH-74

Tested: CQ
Change-Id: I2015dde0bf19bb88ce1acb1867c5eb0db7b875f9