[debugger][x64] PC Jump test
This tests requires writing into the IP in order to be successful.
(See: x64_write_register_test_so.cc).

Also changed the way tests are generated. Now the test package has 3
types of "resources:

1. unit_tests: These correspond to the original debug_agent_tests. This
was changed because one suite mixed unit and integration tests. As
integration tests need to be disabled in certain environments,
separating them into two separate resouces enables different handling.

2 integration_tests: These are tests that exercise Zircon. They require
access to resources (such as shared libraries). They are currently
disabled until there is a way to disable tests for certain environments: DX-909.

3: Generic resources: Utilities applications that are either used by the
tests or useful to replicate certain scenarios manually.

TEST=unit

This is a test designed to require modifying a PC to be a success.

Change-Id: Icc249f8386abfa5c71be538ad794f918643f3df0