[mediaplayer] SimpleStreamSinkImpl fixes
This CL fixes a bug in SimpleStreamSinkImpl (MTWN-215) and addresses
an additional issue with SimpleStreamSinkImpl.

MTWN-215 was caused by a callback lambda accessing |this| after
|this| had been deleted. The bug has been addressed by adding a
shared pointer to |this| to the capture list for the lambda.

SimpleStreamSinkImpl was also using a vector to store added VMOs by
id. This was a mistake given that the ids can be arbitrarily large.
The vector in question has been replaced with an unordered_map.

MTWN-215 #done

TEST: see MTWN-215 repro steps
Change-Id: I57099fe435add8eb7b53fc11bf858a9363be7553