[terminal] Rust terminal uses the terminal model.
The Rust terminal previously included the Alacritty terminal model, but
didn't yet plug in to it. It merely used the canvas class to output a
"hello, world!" string. It still just outputs a "hello, world!" string,
but not it uses the Alacritty terminal model to do it.

Also:
- Fixed a typo, "properies" -> "properties".
- Added a MAINTAINERS file.

Test: Manually tested on Pixelbook in Ermine. Automated tests pending.
This as yet has no consumers, so it has no danger of breaking anything,
as long as it builds, but some basic automated tests are my next
priority.

Change-Id: Icfaac3e335cd8cfe2b9033c5383506fe647f9cb3