[kernel][resource] Initialize resource name in all cases
When a size of zero for the name is passed to zx_resource_create
the syscall could create a resource object whose name was uninitialized
memory on the stack because no copy from userspace was done.

Now we always zero initialization the name buffer.

Test: ran core-tests as well as normal system unit tests, and verified
the output of the resource commands.

ZX-3275 #done

Change-Id: I386e2444f528160c829c57cd5e5e63b687717562