[netstack] Use optional field for the gateway in RouteTableEntry
This changes the FIDL RouteTableEntry field for the gateway to an optional
field and adds code to handle this properly.

This broke recently when we moved to IP Addresses implemented as a union in
FIDL. Before that we used a struct that had an enum NetAddressFamily that
supports "UNSPECIFIED".

This also fixes a couple of nits in ifconfig.

NET-2059 #done

Test: CQ and manual tests creating/deleting/showing routes
Change-Id: Iceed3a4260a0143e9f4bfdc42ad88ba70fa98294