[virtio]: Fake virtio backend
virtio uses a 'backend' abstration to separate device driver
interactions with hardware and interactions with the DDK core.

Introduce a minimal fake backend to allow writing tests of device
drivers. Tests are expected to subclass the Fake backend to intercept
device interactions and provide expected responses to device drivers.

Tested: N/A

Change-Id: I17be1cacca8fd103fbfdbbbec4de9163f0edc727