[mediaplayer] Remove double copy.
This CL removes a copy. Previously, all reads from the upstream source
were copied into a temporary buffer and then copied into
SparseByteBuffer with separate allocations and frees for regions.

Now, the upstream reader writes directly to an address provided by
SlidingBuffer.

This results in a measured framerate increase of 4.9% with --loop and
1.6% with --test-seek, which I assume are more significant than
suggested because the test used SW decode and SW colorspace conversion.

TEST: fx run-test mediaplayer_tests

Change-Id: I9b6665fedaf7e61ddc33b7be4905d25d3d786958