[rust] Create a duplicate of fuchsia-app named fuchsia-component
This will be the new library for building and interacting with services.
Existing applications should transition to use fuchsia-component rather
than fuchsia-app. Rather than just a name change, this will be an
opportunity to build in better support for some of the changes wanted
in fuchsia-app to make the API more flexible and easier to use in a
wider variety of scenarios.

Test: compiled.
Change-Id: I095e58aeb0b6d7aa2e9d937fb0feddd25709986f