[kernel][top] Do not pin bootstrap2 thread
Previously when thread_set_cpu_affinity was called from lk_main no cpus were considered
active, resulting in the thread_set_cpu_affinity call doing nothing and leaving the
bootstrap thread able to run across all cores.

Pinning of the bootstrap2 thread is not needed, and doing so is a leftover from how
bootstrapping used to work. As the call is not needed and not achieving anything this
change removes it for clarity.

Test: Boot zircon on vim2 and NUC

Change-Id: I1708c904223f6f20523502c1911944c04e05c651