[hypervisor] Use spinlock for TrapMap
The TrapMap is used during x86 vmx exit handling where blocking, and potentially thread
switching, is not allowed. As the TrapMap only needs to perform extremely short
operations that do not have further dependencies that might block. We know that TrapTree
operations will not block since they do not perform any allocation as the Trap uses
an intrusive container.

Test: run-test guest_integration_tests

Change-Id: Icd5361cadaa906f0db1b66066436a147b7bc8499