[hypervisor][x86] Prevent blocking whilst holding VMCS
AutoVmcs assumes that whilst it is held (or Invalidate is called) that no other VMCS is
made active on this core. To this end it would disable interrupts to prevent any rescheduling
due to pre-emption from occuring. If another thread would run it could be a VCPU, which
would do a vmptrld, thus changing the active VMCS.

This change extends interrupt disabling to also set blocking_disallowed, which makes it
an error for any blocking operations to be called. Blocking operations include acquiring
any mutexes, as these may block. With this change some of the vmx trap handlers are
changed to call Invalidate if needed, and load any values from the vmcs before doing so.

Test: run-test guset_integration_tests

Change-Id: Icf8aa5988138dc02bdbbd2382269f9ec1ef25a6f