[hypervisor][x86] Check guest interruptibility before interrupt injection
This corrects an observed rare scenario where an interrupt would be injected in the
instruction after an STI, which would result in the VM exit "ENTRY_FAILURE_GUEST_STATE".

Closures are used to provide readability and allowing the vmcs->Read calls to be
lazy preventing them from always being done. They will still be inlined by the compiler
though.

Test: run-test guest_integration_tests

Change-Id: I0c90e2be9b6ce919e8f6c4c7766f30c2ab661668