[guest][vmm] Ensure correct memory orderings when manipulating virtio queues
The virtio specification (3.2.1 Supplying Buffers to The Device) details the requirements
on a driver of the order that ring updates must become visible to the device. Specifically
it must perform any required memory barriers to ensure that the device can see the updated
descriptor table and available ring *before* it could possibly see the updated idx.

Similarly it is the responsibility of the device to ensure that it does not look at the
descriptor table or available ring prior to observing the changed idx. Previously this
was not enforced and it was possible for the CPU to fetch the descriptor and available
rings prior to idx. Even though the accesses were done through volatile pointers this does
not prevent the CPU from re-ordering reads.

Using an atomic load with an ACQUIRE barrier ensures that neither the compiler nor the
hardware can move any loads to before the load of idx.

Similarly when putting descriptor chains into the used ring we use a RELEASE barrier to
ensure the CPU (ARM only as x86 would not reorder this due to TSO) does not re-order
the writes.

Using the acquire and release barriers subsumes the use of volatile as in preventing
loads and stores from moving past them they also force all non stack values to be
reloaded.

Test: run-test guest_integration_tests

Change-Id: I22a169c95de04f07c61dcb4111b36de43905166f