[iquery] Bug workaround for hang issue.
This is a temporary workaround for ZX-3284, which is causing iquery
operations to hang occasionally.

The problem is that components who are resumed from a suspended state
occasionally miss events waiting on a channel. This change adds a new
FilePoker class that takes the path of a directory that will soon be opened.
If the operation to open that file happens to hang, a thread wrapped by
FilePoker will attempt to open the directory, which unsticks the
process.

This hack uncovered a possible additional problem with thread
suspension, but it is far more rare than the original issue.

ZX-3284: #comment
CF-255: #done

Change-Id: I74263679f19446c11e0999a5e522a21fe43c1e71