[netstack] Dereference to prevent allocation
Aliasing &b and &a.iobuf (i.e. the pointers to the slice headers) was
fooling the compiler into thinking the memory they pointed to escaped.
We don't need pointers here, so dereferencing into the stack is
perfectly OK.

Tested that `go build -gcflags=-m` was previously reporting that &b and
&a.iobuf escaped, but now reports that they do not.

Test: go build -gcflags=-m
Change-Id: Iea5430f04c44ed0ec584e9ebfd73ed76c5b742de