[qemu][botanist] Abstract out QEMU invocation generation
This change introduces
(1) QEMUConfig, a high-level configuration interface for fuchsia in
QEMU, and
(2) qemu.CeateInvocation(), a function to generate the invocation from it.

Moreover, this change removes our qcow dependency, which is not needed.

A further change will introduce a GN action that will invoke a simple
tool (wrapping CreateInvocation) to produce a QEMU invocation
file(/script?) so that the configurate used by infrastructure may be
shared with developers, fx, and the SDK.

Test: ran `botanist qemu` locally.

Change-Id: Ib62b33896db90b54ef108afe93d0b49302fe1c53