[biscotti] Default to newer virtio-net device.
The legacy virtio-net device was able to connect to the internet, but
causes a lot of flakey behavior. The new net device enables guest<->host
communication but is not yet WAN routable.

For now lack of internet access is not critical as long as a stateful
partition is copied over so that the VM can skip the container
initialization.

Test: tiles_ctl add linux://chromium
Change-Id: I4cfab294c97b34d82cb95e73294a8411cffea4a5