[devmgr] Pipelined open for device_add.
This allows a client to open a connection to the device
when adding a device, instead of having to wait for the device
appear in the filesystem and open it.

If the device is added with DEVICE_ADD_MUST_ISOLATE,
they will get connected to the proxy device instead.

If the device is added with DEVICE_ADD_INVISIBLE,
they will not be be connected until device_make_visible is called.

ZX-2956 #comment

TEST= I wrote a small device and proxy and verified the channel
worked.

Change-Id: I5e1eed1eedaa6be65d0f0053c4c68dee0ead3058