[sessionmgr] Delay initializing services until a SessionShell comes up and connects.
Currently when the sessionmgr starts it initializes all the services the
session shell needs before it launches the session shell.  This causes a
significant delay to the session shell starting as we need to launch
numerous components before it gets to session shell.

This change makes session shell start before any other services and the
other services are started upon session shell first asking for them.

Test: Logged in and gathered metrics with traceutil to see the order of things being launched.

Change-Id: I42a654b874cbc910957eba5975e6befc28b68d03