[debugger] Separate out names when formatting values.
This separates out the name of the tree node used by FormatValue so the
final stringification of the "name = " part happens at the end. This is
a first step of a larger refactor that will eventually bring the
OutputNode into its own file.

This is described in more detail in the comment added in format_value.h

There should be no user-visible behavior change.

TEST=unit

Change-Id: I1724973dba877bb292f36b27613097eaf0d46d91