[debugger] Improve OutputBuffer.
Removes the deprecated OutputBuffer::WithContents helper functions in
preference to just using the constructors.

Removes OutputBuffer.Set*Color functions in preference to using the new
color constructor and append functions. The Set*Color functions are
surprising because they override any colors or formatting in previously
output text.

Add foreground and background color parameters to the constructor and
Append functions on OutputBuffer. These are defaulted to "kDefault" so
they come after the string.

TEST=unit,manual

Change-Id: I5c21c8064dd8d544706986d9b9b07c2efa228b8b