[usb-request] Add private storage support
Added a template parameter to usb::Request and usb::UnownedRequest to
allow additional private storage for a driver.

An unfortunately side-effect of this is that a template parameter
needs to be specified when no additional private storage is required
(I don't think default template parameters play well with CRTP).

Tested: `runtests -t usb-unittest`

Change-Id: I46ee50f8904992ecf466d71552abcf6292146758