[abigen] Fix syscall generation for one argument and zero returns
The computation of how many C arguments to include would underflow in
the case that there were no abigen-level return values. This was not
noticed, as the value was only compared to be greater than zero, and
all existing syscalls with no return values have at least two
abigen-level parameters.

This was discovered by an experiment to make _zx_handle_close return
void. As it then has one parameter and no return values, the underflow
was observable.

Test: manual inspection of out/.../syscalls/definitions.h
Change-Id: I25a20a17fb939dbf069ae5e6ad132d3c70ccec99