[rust] Guest trap definitions in fuchsia-zircon-sys
The previous guest trap definitions had diverged both in their accuracy with the kernel
definitions, and from how other port packets in the fuchsia-zircon-sys were defined.
This change updates all the definitions to match the current kernel and moves the style
of the definitions to be in line with other port definitions.

Test: build garnet. run-test fuchsia_zircon_lib_test
Change-Id: I90b4e9eff00c6ed162d5e81d53e1dfb399954073