[kernel][x86][topology] Generate SystemTopology on IA x86.
For intel processors, this will create a flat topology to populate the
Global SystemTopology class. It uses ACPI and CPUID information to do
this.

We don't yet use the Global SystemTopology class on x86, first we must
generate the matching code for AMD x86.

This also moves test data for the cpuid and acpi_tables classes to
headers so it can be shared with these tests.

Test: Added unit test case that parses real, but static, data for the HP z840.
Issue: ZX-3068
Change-Id: I55c84e18ef522d49e0b58d80798334c9b1680a36